requirements.txt
pandas==1.5.0
streamlit==1.19.0
PyPDF2==2.12.1
gspread==5.8.0
google-auth==2.17.3
